+++ Changelog.md
@@ -4,6 +4,9 @@
 
 ### Ignition Gazebo 3.0.0
 
+1. Add Marker Manager
+    * [Pull Request 442](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/442)
+
 1. Parse material emissive map, bump to msgs5 and transport8
     * [Pull Request 447](https://bitbucket.org/ignitionrobotics/ign-gazebo/pull-requests/447)